requirements.txt
aiogram~=3.16.0
google-genai
PyYAML~=6.0.1
pillow~=10.3.0
Pyrogram~=2.0.106